requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.9.1
asn1crypto==1.5.1
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.5.4
cryptography==46.0.2
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-environ==0.4.5
django-health-check==3.18.1
django-htmx==1.23.2
django-redis==5.4.0
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
factory-boy==3.3.0
Faker==37.8.0
flake8==7.3.0
freetype-py==2.5.1
ghp-import==2.1.0
gunicorn==23.0.0
honcho==2.0.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
isort==6.1.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.26
mkdocs-material-extensions==1.3.1
mypy_extensions==1.1.0
numpy==2.1.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
paginate==0.5.7
pandas==2.2.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycairo==1.28.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pymdown-extensions==10.16.1
pypdf==6.1.1
pytest==8.2.2
pytest-django==4.8.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
redis==6.4.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.4
rich==14.1.0
rlPyCairo==0.4.0
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.3.1
six==1.17.0
soupsieve==2.5
sqlparse==0.5.3
svglib==1.6.0
tailwind==3.1.5b0
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
